+++ src/collective/cover/tiles/list.py
@@ -18,7 +18,6 @@ from plone.uuid.interfaces import IUUID
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope import schema
 from zope.component import queryUtility
-from zope.deprecation.deprecation import deprecate
 from zope.interface import implementer
 from zope.schema import getFieldsInOrder
 
@@ -219,10 +218,6 @@ class ListTile(PersistentCoverTile):
 
         old_data['uuids'] = uuids_dict
         data_mgr.set(old_data)
-
-    @deprecate('Use replace_with_uuids method instead.')
-    def replace_with_objects(self, uuids):
-        self.replace_with_uuids(uuids)
 
     def replace_with_uuids(self, uuids):
         """ Replaces the whole list of items with a new list of items